@@ -1,6 +1,6 @@
 <Project Sdk="Microsoft.NET.Sdk">
   <PropertyGroup>
-    <TargetFrameworks>net6.0;netstandard2.1</TargetFrameworks>
+    <TargetFrameworks>net7.0;net6.0;netstandard2.1</TargetFrameworks>
     <RootNamespace>NCoreUtils.Images</RootNamespace>
     <!-- NuGet -->
     <PackageId>NCoreUtils.Images.Abstractions</PackageId>
@@ -21,11 +21,13 @@
     <!-- <GenerateDocumentationFile>true</GenerateDocumentationFile> -->
     <TreatWarningsAsErrors>true</TreatWarningsAsErrors>
   </PropertyGroup>
-  <ItemGroup Condition=" '$(TargetFramework)' == 'net6.0' ">
+  <ItemGroup Condition=" '$(TargetFramework)' != 'netstandard2.1' ">
     <Compile Remove="CodeAnalysis\**" />
+    <Compile Remove="CompilerServices\**" />
   </ItemGroup>
   <ItemGroup>
-    <PackageReference Include="NCoreUtils.Extensions.Memory" Version="3.0.0.5" />
-    <PackageReference Include="NCoreUtils.IO.Abstractions" Version="6.0.1" />
+    <PackageReference Include="NCoreUtils.Extensions.Memory" Version="7.0.0" />
+    <PackageReference Include="NCoreUtils.IO.Abstractions" Version="7.0.0" />
+    <PackageReference Include="NCoreUtils.Resources.Abstractions" Version="7.0.0" />
   </ItemGroup>
 </Project>
