@@ -1,52 +1,51 @@
-﻿namespace NCoreUtils.Images.Internal
+﻿namespace NCoreUtils.Images.Internal;
+
+public enum WaterMarkGravity
 {
-    public enum WaterMarkGravity
-    {
-        //
-        // Summary:
-        //     Undefined.
-        Undefined = 0,
-        //
-        // Summary:
-        //     Forget.
+    //
+    // Summary:
+    //     Undefined.
+    Undefined = 0,
+    //
+    // Summary:
+    //     Forget.
 #pragma warning disable CA1069
-        Forget = 0,
+    Forget = 0,
 #pragma warning restore CA1069
-        //
-        // Summary:
-        //     Northwest.
-        Northwest = 1,
-        //
-        // Summary:
-        //     North.
-        North = 2,
-        //
-        // Summary:
-        //     Northeast.
-        Northeast = 3,
-        //
-        // Summary:
-        //     West.
-        West = 4,
-        //
-        // Summary:
-        //     Center.
-        Center = 5,
-        //
-        // Summary:
-        //     East.
-        East = 6,
-        //
-        // Summary:
-        //     Southwest.
-        Southwest = 7,
-        //
-        // Summary:
-        //     South.
-        South = 8,
-        //
-        // Summary:
-        //     Southeast.
-        Southeast = 9
-    }
+    //
+    // Summary:
+    //     Northwest.
+    Northwest = 1,
+    //
+    // Summary:
+    //     North.
+    North = 2,
+    //
+    // Summary:
+    //     Northeast.
+    Northeast = 3,
+    //
+    // Summary:
+    //     West.
+    West = 4,
+    //
+    // Summary:
+    //     Center.
+    Center = 5,
+    //
+    // Summary:
+    //     East.
+    East = 6,
+    //
+    // Summary:
+    //     Southwest.
+    Southwest = 7,
+    //
+    // Summary:
+    //     South.
+    South = 8,
+    //
+    // Summary:
+    //     Southeast.
+    Southeast = 9
 }
