@@ -1,22 +1,5 @@
 using System.Numerics;
 
-namespace NCoreUtils.Images.ImageMagick
-{
-#if NETSTANDARD2_1
-    public struct UsageData
-    {
-        public BigInteger Total { get; }
+namespace NCoreUtils.Images.ImageMagick;
 
-        public BigInteger Used { get; }
-
-        public UsageData(BigInteger Total, BigInteger Used)
-        {
-            this.Total = Total;
-            this.Used = Used;
-        }
-    }
-
-#else
-    public readonly record struct UsageData(BigInteger Total, BigInteger Used);
-#endif
-}
+public readonly record struct UsageData(BigInteger Total, BigInteger Used);
