@@ -1,19 +1,18 @@
 using System.Collections.Generic;
 using NCoreUtils.Images.Internal;
 
-namespace NCoreUtils.Images
+namespace NCoreUtils.Images;
+
+public class ResizerCollectionBuilder
 {
-    public class ResizerCollectionBuilder
+    readonly Dictionary<string, IResizerFactory> _factories = new();
+
+    public ResizerCollectionBuilder Add(string name, IResizerFactory factory)
     {
-        readonly Dictionary<string, IResizerFactory> _factories = new();
-
-        public ResizerCollectionBuilder Add(string name, IResizerFactory factory)
-        {
-            _factories.Add(name, factory);
-            return this;
-        }
-
-        public ResizerCollection Build()
-            => new(_factories);
+        _factories.Add(name, factory);
+        return this;
     }
+
+    public ResizerCollection Build()
+        => new(_factories);
 }
