@@ -1,6 +1,7 @@
-using NCoreUtils.Memory;
-
 namespace NCoreUtils.Images.Internal
 {
-    public interface IFilter : IEmplaceable<IFilter> { }
+    public interface IFilter
+    {
+        int GetEmplaceBufferSize();
+    }
 }
