@@ -27,14 +27,29 @@ namespace NCoreUtils.Images.Logging
             Optimize = optimize;
         }
 
+        bool IEmplaceable<ResizingImageEntry>.TryGetEmplaceBufferSize(out int minimumBufferSize)
+        {
+            minimumBufferSize = 52 + 12 + 13 + 21 + 5 + ImageType.Length;
+            if (!IsExplicit)
+            {
+                minimumBufferSize += 11;
+            }
+            return true;
+        }
+
+#if NET6_0_OR_GREATER
+        string IFormattable.ToString(string? format, System.IFormatProvider? formatProvider)
+            => ToString();
+#else
         public int Emplace(Span<char> span)
         {
             if (TryEmplace(span, out var used))
             {
                 return used;
             }
-            throw new ArgumentException("Insufficient buffer size.", nameof(span));
+            throw new InsufficientBufferSizeException(span);
         }
+#endif
 
         public bool TryEmplace(Span<char> span, out int used)
         {
@@ -58,29 +73,7 @@ namespace NCoreUtils.Images.Logging
             return false;
         }
 
-        [MethodImpl(MethodImplOptions.NoInlining)]
-        private bool TryToStringOnStack([NotNullWhen(true)] out string? result)
-        {
-            Span<char> buffer = stackalloc char[4 * 1024];
-            if (TryEmplace(buffer, out var size))
-            {
-                result = buffer[..size].ToString();
-                return true;
-            }
-            result = default;
-            return false;
-        }
-
         public override string ToString()
-        {
-            if (TryToStringOnStack(out var result))
-            {
-                return result;
-            }
-            using var memoryBuffer = MemoryPool<char>.Shared.Rent(32 * 1024);
-            var buffer = memoryBuffer.Memory.Span;
-            var size = Emplace(buffer);
-            return buffer[..size].ToString();
-        }
+            => Emplacer.ToStringUsingArrayPool(this);
     }
 }
